+++ src/onetick/__init__.py
@@ -1,4 +1,3 @@
-import sys
 
 
 def __search_main_one_tick_dir():
@@ -16,7 +15,7 @@ def __search_main_one_tick_dir_catch(show_warnings=True):  # noqa: C901
     import sysconfig
     from pathlib import Path
 
-    if os.environ.get('OTP_SKIP_OTQ_VALIDATION', False):
+    if os.environ.get('OTP_SKIP_OTQ_VALIDATION'):
         try:
             import onetick_stubs  # noqa
         except ImportError:
@@ -95,7 +94,22 @@ def __search_main_one_tick_dir_catch(show_warnings=True):  # noqa: C901
     return (ot_bin_path, ot_python_path, ot_numpy_path)
 
 
-for directory in __search_main_one_tick_dir() or []:
-    sys.path.append(str(directory))
+# in this __init__.py file we only need to modify sys.path to be able to import any onetick namespace library
+
+try:
+    # this block works if we are in a onetick wheel build
+    # (onetick binaries and onetick.query are installed with pip)
+    # if we don't have import error, then sys.path was already modified in env.py from onetick wheel
+    from . import env  # type: ignore
+    from .__version__ import __version__  # noqa
+    __build__ = ''
+except ImportError:
+    # otherwise we are in a regular onetick-py installation and
+    # we need to search for MAIN_ONE_TICK_DIR env variable and modify sys.path ourselves
+    import sys
+    sys.path.extend(
+        str(directory) for directory in __search_main_one_tick_dir() or []
+    )
+    del sys
 
 __path__ = __import__("pkgutil").extend_path(__path__, __name__)