+++ tests/core/sources/test_dataframe.py
@@ -1,4 +1,5 @@
 from io import StringIO
+from textwrap import dedent
 
 import pandas as pd
 import pytest
@@ -61,7 +62,7 @@ class TestReadFromDataFrame:
     @pytest.mark.parametrize('timestamp_column', [
         'TIME', 'TiMe', 'Time', 'TIMESTAMP', 'Timestamp',
     ])
-    def test_timestamp_column_auto(self, f_session, timestamp_column):
+    def test_timestamp_column_auto(self, session, timestamp_column):
         dataframe = self.get_test_data(timestamp_column=timestamp_column, add_special_values=False)
 
         src_schema = {
@@ -92,7 +93,7 @@ class TestReadFromDataFrame:
             pd.Timestamp('2024-01-01 12:00:03.250')
         ]
 
-    def test_timestamp_column(self, f_session):
+    def test_timestamp_column(self, session):
         timestamp_column = 'TEST'
         dataframe = self.get_test_data(timestamp_column=timestamp_column)
 
@@ -111,7 +112,7 @@ class TestReadFromDataFrame:
         assert len(compat) == 5
         assert set(compat.columns) == result_schema
 
-    def test_empty_timestamp_column(self, f_session):
+    def test_empty_timestamp_column(self, session):
         timestamp_column = 'TEST'
         dataframe = self.get_test_data(timestamp_column=timestamp_column)
 
@@ -131,7 +132,7 @@ class TestReadFromDataFrame:
         assert len(compat) == 5
         assert set(compat.columns) == result_schema
 
-    def test_datetime_timestamp_column(self, f_session):
+    def test_datetime_timestamp_column(self, session):
         timestamp_column = 'Timestamp'
         dataframe = self.get_test_data(timestamp_column=timestamp_column)
         dataframe[timestamp_column] = dataframe[timestamp_column].apply(pd.to_datetime)
@@ -151,7 +152,7 @@ class TestReadFromDataFrame:
         assert len(compat) == 5
         assert set(compat.columns) == result_schema
 
-    def test_timestamp_column_without_autodetect(self, f_session):
+    def test_timestamp_column_without_autodetect(self, session):
         timestamp_column = 'Timestamp'
         dataframe = self.get_test_data(timestamp_column=timestamp_column)
 
@@ -175,7 +176,7 @@ class TestReadFromDataFrame:
         assert len(compat) == 5
         assert set(compat.columns) == result_schema
 
-    def test_special_values_compat(self, f_session):
+    def test_special_values_compat(self, session):
         dataframe = self.get_test_data(timestamp_column='TIME', add_special_values=True)
 
         src_schema = {
@@ -193,7 +194,7 @@ class TestReadFromDataFrame:
         assert len(compat) == 6
         assert set(compat.columns) == result_schema
 
-    def test_symbol_name_field(self, f_session):
+    def test_symbol_name_field(self, session):
         dataframe = self.get_test_data(timestamp_column='Timestamp', symbol_name_field='TEST', symbol_name='AAPL')
 
         src_schema = {
@@ -209,7 +210,7 @@ class TestReadFromDataFrame:
         assert len(compat) == 5
         assert set(compat.columns) == result_schema
 
-    def test_exceptions(self):
+    def test_exceptions(self, session):
         df = self.get_test_data(timestamp_column='Timestamp')
 
         with pytest.raises(ValueError, match='DataFrame should be passed'):
@@ -244,3 +245,70 @@ class TestReadFromDataFrame:
 
         with pytest.raises(ValueError, match='not in dataframe'):
             otp.ReadFromDataFrame(df, timestamp_column='MISSING', symbol='AAPL')
+
+    def test_run(self, session):
+        input_df = self.get_test_data()
+        src = otp.ReadFromDataFrame(input_df)
+        assert src.schema == {'ID': int, 'SIDE': str, 'PRICE': float, 'FLOAT': float, 'SYMBOL_NAME': str}
+        df = otp.run(src, date=otp.dt(2024, 1, 1))
+        assert list(df['Time']) == [otp.dt(2024, 1, 1, 12, 0, 0, 1000),
+                                    otp.dt(2024, 1, 1, 12, 0, 2),
+                                    otp.dt(2024, 1, 1, 12, 0, 2, 500000),
+                                    otp.dt(2024, 1, 1, 12, 0, 3, 100000),
+                                    otp.dt(2024, 1, 1, 12, 0, 3, 250000)]
+        assert list(df['ID']) == [1, 1, 2, 2, 3]
+        assert list(df['PRICE']) == [50.05, 50.05, 49.95, 49.98, 50.02]
+        assert list(df['FLOAT']) == list(df['PRICE'])
+        assert list(df['SYMBOL_NAME']) == ['AAPL', 'AAPL', 'AAPL', 'AAPL', 'AAPL']
+
+    def test_copy(self, session):
+        # PY-1484  (reproduces only with pyarrow installed)
+        input_df = self.get_test_data()
+        src = otp.ReadFromDataFrame(input_df)
+        src = src.copy()
+        assert src.schema == {'ID': int, 'SIDE': str, 'PRICE': float, 'FLOAT': float, 'SYMBOL_NAME': str}
+        df = otp.run(src, date=otp.dt(2024, 1, 1))
+        assert list(df['Time']) == [otp.dt(2024, 1, 1, 12, 0, 0, 1000),
+                                    otp.dt(2024, 1, 1, 12, 0, 2),
+                                    otp.dt(2024, 1, 1, 12, 0, 2, 500000),
+                                    otp.dt(2024, 1, 1, 12, 0, 3, 100000),
+                                    otp.dt(2024, 1, 1, 12, 0, 3, 250000)]
+        assert list(df['ID']) == [1, 1, 2, 2, 3]
+        assert list(df['PRICE']) == [50.05, 50.05, 49.95, 49.98, 50.02]
+        assert list(df['FLOAT']) == list(df['PRICE'])
+        assert list(df['SYMBOL_NAME']) == ['AAPL', 'AAPL', 'AAPL', 'AAPL', 'AAPL']
+
+    def test_update(self, session):
+        # PY-1484 (reproduces only with pyarrow installed)
+        csv_input = dedent(
+            """
+            TIMESTAMP,STATE,ID,SIDE,PRICE,ORIG_QTY,LEAVES_QTY,FILL_QTY,FILL_PRICE
+            2024-01-03 14:33:00.000,N,1,BUY,50.05,100,100,0,nan
+            2024-01-03 14:33:02.000,PF,1,BUY,50.05,100,50,50,50.05
+            2024-01-03 14:33:32.000,F,1,BUY,50.05,100,0,50,50.05
+            2024-01-03 14:34:42.000,N,2,SELL,50.30,100,0,0,nan
+            2024-01-03 14:34:52.000,C,2,SELL,50.30,100,0,0,nan
+            2024-01-03 14:36:00.000,N,3,BUY,49.98,100,100,0,nan
+            2024-01-03 14:36:02.000,PF,3,BUY,49.98,100,80,20,49.97
+            2024-01-03 14:36:03.000,PF,3,BUY,49.98,100,60,20,49.97
+            2024-01-03 14:36:03.250,PF,3,BUY,49.98,100,40,20,49.97
+            2024-01-03 14:36:03.260,PF,3,BUY,49.98,100,20,20,49.97
+            2024-01-03 14:36:04.000,PF,3,BUY,49.98,100,10,10,49.97
+            2024-01-03 14:36:12.000,F,3,BUY,49.98,100,0,10,49.95
+            """
+        )
+        # CSV with TIMESTAMP field set to %Y%m%d%H%M%S.%f format in UTC
+        df = pd.read_csv(StringIO(csv_input))
+        df['TIMESTAMP'] = pd.to_datetime(df['TIMESTAMP'])
+
+        # Load DataFrame as a source, specifying the symbol
+        orders = otp.ReadFromDataFrame(df, symbol='AAPL', db='LOCAL')
+        orders['NEW_QTY'] = orders['ORIG_QTY']
+        orders = orders.update(if_set={orders['NEW_QTY']: orders['LEAVES_QTY']}, where=(orders['STATE'] == 'PF'))
+
+        # Run the query for a specific date
+        result = otp.run(orders, date=otp.date(2024, 1, 3))
+        assert list(result['STATE']) == ['N', 'PF', 'F', 'N', 'C', 'N', 'PF', 'PF', 'PF', 'PF', 'PF', 'F']
+        assert list(result['ORIG_QTY']) == [100] * 12
+        assert list(result['LEAVES_QTY']) == [100, 50, 0, 0, 0, 100, 80, 60, 40, 20, 10, 0]
+        assert list(result['NEW_QTY']) == [100, 50, 100, 100, 100, 100, 80, 60, 40, 20, 10, 100]